ROFO-29 패키지 구조 정리
* 패키지 구조를 도메인별로 정리하였습니다.
