+++ src/main/kotlin/kr/weit/roadyfoody/common/exception/CustomExceptions.kt
@@ -1,0 +1,5 @@
+package kr.weit.roadyfoody.common.exception
+
+open class BaseException(val errorCode: ErrorCode, message: String) : RuntimeException(message)
+
+class UserNotFoundException(message: String) : BaseException(ErrorCode.NO_SUCH_ELEMENT, message)
