--- src/main/kotlin/kr/weit/roadyfoody/config/JpaAuditingConfig.kt
@@ -1,8 +1,0 @@
-package kr.weit.roadyfoody.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@Configuration
-@EnableJpaAuditing
-class JpaAuditingConfig
