--- src/main/kotlin/kr/weit/roadyfoody/controller/HealthController.kt
@@ -1,23 +1,0 @@
-package kr.weit.roadyfoody.controller
-
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.context.annotation.Profile
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@Profile("!test")
-@RestController
-class HealthController {
-    @GetMapping("/health")
-    fun health(): HealthResponse {
-        return HealthResponse(status = HttpServletResponse.SC_OK)
-    }
-
-    @GetMapping("/ready")
-    fun ready() {
-    }
-}
-
-data class HealthResponse(
-    val status: Int,
-)
