--- src/main/kotlin/kr/weit/roadyfoody/domain/BaseTimeEntity.kt
@@ -1,26 +1,0 @@
-package kr.weit.roadyfoody.domain
-
-import jakarta.persistence.Column
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.MappedSuperclass
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-abstract class BaseTimeEntity {
-    @CreatedDate
-    @Column(name = "created_datetime", nullable = false, updatable = false)
-    lateinit var createdDateTime: LocalDateTime
-        protected set
-}
-
-@MappedSuperclass
-abstract class BaseModifiableEntity : BaseTimeEntity() {
-    @LastModifiedDate
-    @Column(name = "updated_datetime", nullable = false)
-    lateinit var updatedDateTime: LocalDateTime
-        protected set
-}
