+++ src/main/kotlin/kr/weit/roadyfoody/global/config/JpaAuditingConfig.kt
@@ -1,0 +1,8 @@
+package kr.weit.roadyfoody.global.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+
+@Configuration
+@EnableJpaAuditing
+class JpaAuditingConfig
