+++ src/main/kotlin/kr/weit/roadyfoody/global/swagger/v1/SwaggerTag.kt
@@ -1,0 +1,6 @@
+package kr.weit.roadyfoody.global.swagger.v1
+
+// SwaggerTag는 Swagger에서 사용할 태그를 정의하는 object입니다.
+object SwaggerTag {
+    const val TEST: String = "Z. TEST용 API"
+}
