--- src/main/kotlin/kr/weit/roadyfoody/security/filter/MockPassFilter.kt
@@ -1,32 +1,0 @@
-package kr.weit.roadyfoody.security.filter
-
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.security.core.userdetails.User
-import org.springframework.web.filter.OncePerRequestFilter
-
-class MockPassFilter : OncePerRequestFilter() {
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain,
-    ) {
-        if (request.headerNames.toList().contains("userid")) {
-            val userDetails =
-                User(
-                    "ForDev",
-                    "",
-                    listOf(
-                        SimpleGrantedAuthority("ROLE_USER"),
-                    ),
-                )
-            SecurityContextHolder.getContext().authentication =
-                UsernamePasswordAuthenticationToken(userDetails, null, userDetails.authorities)
-        }
-        filterChain.doFilter(request, response)
-    }
-}
