--- src/main/kotlin/kr/weit/roadyfoody/support/exception/ErrorCode.kt
@@ -1,14 +1,0 @@
-package kr.weit.roadyfoody.support.exception
-
-import org.springframework.http.HttpStatus
-
-enum class ErrorCode(val httpStatus: HttpStatus, val code: Int, val errorMessage: String) {
-    INVALID_REQUEST(HttpStatus.BAD_REQUEST, -10000, "Invalid request"),
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, -10001, "Unauthorized"),
-    FORBIDDEN(HttpStatus.FORBIDDEN, -10002, "Forbidden"),
-    NO_SUCH_ELEMENT(HttpStatus.NOT_FOUND, -10003, "No such element"),
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, -10004, "Internal server error"),
-    EXIST_RESOURCE(HttpStatus.CONFLICT, -10005, "Exist resource"),
-    NOT_FOUND_DEFAULT_RESOURCE(HttpStatus.INTERNAL_SERVER_ERROR, -10007, "Not found default resource"),
-    PAYLOAD_TOO_LARGE(HttpStatus.PAYLOAD_TOO_LARGE, -10008, "Payload too large"),
-}
